split functionality in multiple classes
